requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
boto3==1.34.34
python-multipart==0.0.9
jinja2==3.1.3
itsdangerous==2.1.2
starlette==0.36.3
